requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.2
base58==2.1.1
bitarray==2.6.0
black==23.3.0
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
eth2spec==1.3.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.1
lru-dict==1.1.8
marko==1.0.2
milagro-bls-binding==1.9.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==1.0.0
netaddr==0.8.0
packaging==23.1
parsimonious==0.9.0
pathspec==0.11.1
platformdirs==3.5.0
protobuf==4.22.3
py-ecc==6.0.0
py_arkworks_bls12381==0.3.4
pycodestyle==2.10.0
pycryptodome==3.15.0
pyrsistent==0.16.1
regex==2023.3.23
remerkleable==0.1.27
requests==2.28.1
rlp==3.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sortedcontainers==2.4.0
tomli==2.0.1
toolz==0.12.0
trie==2.0.2
urllib3==1.26.13
varint==1.0.2
websockets==11.0.2
yarl==1.8.1
